added: DetailedAnimeSummary component
